+++ Network/API/Telegram/Bot/Object/Update/Message/Content/Info/Location.hs
@@ -7,8 +7,7 @@ import "base" Data.Functor ((<$>))
 import "base" GHC.Float (Float)
 import "base" Text.Show (Show)
 
-data Location = Location Float Float
-	deriving Show
+data Location = Location Float Float deriving Show
 
 instance FromJSON Location where
 	parseJSON = withObject "Location" $ \v -> Location